+++ src/main/java/faang/school/postservice/exception/post/PostDeletedException.java
@@ -1,6 +1,6 @@
 package faang.school.postservice.exception.post;
 
-import faang.school.postservice.exception.exceptionmessages.PostServiceExceptionMessage;
+import faang.school.postservice.exception.messages.PostServiceExceptionMessage;
 
 public class PostDeletedException extends PostServiceException {
     public PostDeletedException(PostServiceExceptionMessage message, Object... args) {
