+++ src/main/java/faang/school/postservice/exception/post/PostServiceException.java
@@ -1,10 +1,10 @@
 package faang.school.postservice.exception.post;
 
 import faang.school.postservice.exception.BaseRuntimeException;
-import faang.school.postservice.exception.exceptionmessages.PostServiceExceptionMessage;
+import faang.school.postservice.exception.messages.PostServiceExceptionMessage;
 
 public class PostServiceException extends BaseRuntimeException {
     public PostServiceException(PostServiceExceptionMessage message, Object... args) {
-        super(message.getMsg(), args);
+        super(message.getMessage(), args);
     }
 }
