+++ src/main/java/faang/school/postservice/exception/post/UnexistentPostPublisher.java
@@ -1,6 +1,6 @@
 package faang.school.postservice.exception.post;
 
-import faang.school.postservice.exception.exceptionmessages.PostServiceExceptionMessage;
+import faang.school.postservice.exception.messages.PostServiceExceptionMessage;
 
 public class UnexistentPostPublisher extends PostServiceException{
     public UnexistentPostPublisher(PostServiceExceptionMessage message, Object... args) {
