+++ src/main/java/faang/school/postservice/exception/validation/DataValidationException.java
@@ -1,7 +1,7 @@
 package faang.school.postservice.exception.validation;
 
 import faang.school.postservice.exception.BaseRuntimeException;
-import faang.school.postservice.exception.exceptionmessages.ValidationExceptionMessage;
+import faang.school.postservice.exception.messages.ValidationExceptionMessage;
 
 public class DataValidationException extends BaseRuntimeException {
     public DataValidationException(ValidationExceptionMessage validationMsg, Object... args) {
